requirements.txt
Flask
httpx
pytz
python-dotenv
stravalib==1.6.0